# openapi.yaml
openapi: 3.1.0
info:
  title: CivicLens API
  version: 1.0.0
  description: >
    AI-assisted public-service navigation API. Eligibility determinations are
    always produced by the deterministic rule engine (see
    docs/ai/eligibility-engine.md); the assistant endpoints are
    retrieval-grounded and always cite sources.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /applications:
    get:
      summary: List the current citizen's applications
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        "200":
          description: Paginated applications
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationPage' }
    post:
      summary: Start an application for a scheme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheme_id]
              properties:
                scheme_id: { type: string, format: uuid }
      responses:
        "201":
          description: Draft application created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        "409":
          description: Citizen is not eligible / likely_eligible for this scheme
  /applications/{application_id}:
    get:
      summary: Get application detail including status history
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationDetail' }
        "404": { $ref: '#/components/responses/NotFound' }
  /applications/{application_id}/submit:
    post:
      summary: Submit a completed application (validates completeness)
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200": { description: Submitted }
        "422":
          description: Application incomplete (missing documents/fields)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /applications/{application_id}/withdraw:
    post:
      summary: Withdraw an application prior to a terminal status
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200": { description: Withdrawn }
  /applications/{application_id}/checklist:
    get:
      summary: Get application document readiness checklist
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200": { description: Checklist }
  /applications/{application_id}/assign:
    post:
      summary: Assign case worker to application
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200": { description: Assigned }
  /applications/{application_id}/review:
    post:
      summary: Case worker/admin review decision
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200": { description: Reviewed }
  /applications/{application_id}/resolve-action:
    post:
      summary: Citizen resolves requested action
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200": { description: Action resolved }
  /applications/{application_id}/complete:
    post:
      summary: Case worker/admin completes application
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200": { description: Completed }
  /applications/{application_id}/export:
    get:
      summary: Export the application package as PDF
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        "200":
          description: PDF package
          content:
            application/pdf: {}

  # ---------- Notifications ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    ApplicationIdParam:
      name: application_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
...
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        document_type: { type: string }
        status: { type: string, enum: [uploaded, processing, verified, rejected, verification_required, validation_failed, processing_failed] }
        uploaded_at: { type: string, format: date-time }
...
    ApplicationPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Application' }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    Application:
      type: object
      properties:
        id: { type: string, format: uuid }
        scheme_id: { type: string, format: uuid }
        status: { type: string, enum: [draft, ready_for_submission, submission_pending, submission_failed, submitted, under_review, action_required, info_requested, approved, rejected, withdrawn, completed] }
        created_at: { type: string, format: date-time }
        submitted_at: { type: string, format: date-time, nullable: true }
...
    ApplicationDetail:
      allOf:
        - $ref: '#/components/schemas/Application'
        - type: object
          properties:
            status_history:
              type: array
              items:
                type: object
                properties:
                  from_status: { type: string, nullable: true }
                  to_status: { type: string }
                  note: { type: string, nullable: true }
                  created_at: { type: string, format: date-time }
            attached_documents:
              type: array
              items: { $ref: '#/components/schemas/Document' }
...